requirements.txt
Flask==2.3.3
mysql-connector-python==8.3.0
Werkzeug==2.3.7